[REF] LadderFactory 관련 클래스 리팩토링
