+++ src/main/java/laddergame/LadderFactory.java
@@ -3,6 +3,6 @@ package laddergame;
 import laddergame.domain.ladder.Ladder;
 import laddergame.domain.ladder.LadderHeight;
 
-public interface Factory {
+public interface LadderFactory {
     Ladder create(final LadderHeight height, final int width);
 }
