+++ src/main/java/laddergame/domain/ProhibitDuplicateNamesFactory.java
@@ -5,7 +5,7 @@ import laddergame.domain.player.Player;
 import java.util.HashSet;
 import java.util.List;
 
-public abstract class ProhibitDuplicateNamesNamesFactory extends AbstractNamesFactory {
+public abstract class ProhibitDuplicateNamesFactory extends AbstractNamesFactory {
 
 	protected void checkDuplication(List<Player> players) throws IllegalArgumentException {
 		if (new HashSet<>(players).size() != players.size()) {
